RDM-3203 : Fix exporting util functions.
