+++ src/shared/fixture/shared.fixture.ts
@@ -1,7 +1,8 @@
 import { CaseField } from '../domain/definition/case-field.model';
 import { CaseEventTrigger } from '../domain/case-view/case-event-trigger.model';
-import { CaseEventData } from '../domain';
-import { AbstractAppConfig } from '../../app.config';
+import { FieldTypeEnum, WizardPage } from '../domain';
+import createSpyObj = jasmine.createSpyObj;
+import { ShowCondition } from '../conditional-show';
 
 export let createCaseEventTrigger = (id: string,
                                       name: string,
@@ -21,4 +22,29 @@ export let createCaseEventTrigger = (id: string,
   eventTrigger.case_fields = case_fields;
   eventTrigger.can_save_draft = can_save_draft;
   return eventTrigger;
+};
+
+export let aCaseField = (id: string, label: string, type: FieldTypeEnum, display_context: string,
+  show_summary_content_option: number): CaseField => {
+  return {
+    id: id || 'personFirstName',
+    field_type: {
+      id: type.toString() || 'Text',
+      type: type || 'Text'
+    },
+    display_context: display_context || 'OPTIONAL',
+    label: label || 'First name',
+    show_summary_content_option: show_summary_content_option
+  };
+};
+
+export let aWizardPage = (pageId: string, label: string, order: number): WizardPage => {
+  let wp = new WizardPage();
+  wp.id = pageId;
+  wp.label = label;
+  wp.order = order;
+  let parsedShowCondition = createSpyObj<ShowCondition>('parsedShowCondition', ['match']);
+  parsedShowCondition.match.and.returnValue(true);
+  wp.parsedShowCondition = parsedShowCondition;
+  return wp;
 };
